Move second batch of property tests from inside UnusedPrivateMemberSpec
